pyproject.toml
[tool.poetry]
name = "photoai"
version = "0.1.0"
description = "photo scan"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
face-recognition = "^1.3.0"
exifread = "^3.3.0"
face-recognition-models = "^0.3.0"
distribute = "^0.7.3"
matplotlib = "^3.10.1"
pyyaml = "^6.0.2"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
